修改代码

+++ src/main/java/org/jboot/generator/enums/EntityMapping.java
@@ -28,7 +28,7 @@ public enum EntityMapping {
     /**
      * Identifies the generic SQL type {@code BIGINT}.
      */
-    BIGINT(Types.BIGINT, "Long", "BIGINT"),
+    BIGINT(Types.BIGINT, "Long", "INTEGER"),
     /**
      * Identifies the generic SQL type {@code FLOAT}.
      */
